+++ Sources/UI/Shapes/Placard.swift
@@ -46,7 +46,7 @@ public struct Placard: Shape {
     }
 }
 
-@available(iOS 14, macOS 11, tvOS 14, watchOS 7, *)
+@available(iOS 13, macOS 10.15, tvOS 13, watchOS 6, *)
 public struct PlacardShowcaseView: View {
     public init() {}
     public var body: some View {